# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Edge Platform industrial IoT / HMI cloud API.",
        "title": "Edge Platform API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/system/dict-data": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "字典数据"
                ],
                "summary": "新增字典数据",
                "parameters": [
                    {
                        "description": "字典数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dictionary.dataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/dict-data/batch-delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "字典数据"
                ],
                "summary": "批量删除字典数据",
                "parameters": [
                    {
                        "description": "字典数据 ID 列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dictionary.batchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/dict-data/page": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "字典数据"
                ],
                "summary": "字典数据分页",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，1-500",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "字典类型 ID",
                        "name": "dictTypeId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典编码",
                        "name": "dictCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典标签",
                        "name": "dictLabel",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典值",
                        "name": "dictValue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/dict-data/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "字典数据"
                ],
                "summary": "字典数据详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典数据 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "字典数据"
                ],
                "summary": "修改字典数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典数据 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "字典数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dictionary.dataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "字典数据"
                ],
                "summary": "删除字典数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典数据 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/dict-type": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "字典类型"
                ],
                "summary": "新增字典类型",
                "parameters": [
                    {
                        "description": "字典类型",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dictionary.typeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/dict-type/batch-delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "字典类型"
                ],
                "summary": "批量删除字典类型",
                "parameters": [
                    {
                        "description": "字典类型 ID 列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dictionary.batchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/dict-type/page": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "字典类型"
                ],
                "summary": "字典类型分页",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，1-500",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典名称",
                        "name": "dictName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典编码",
                        "name": "dictCode",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态：0 或 1",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/system/dict-type/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "字典类型"
                ],
                "summary": "修改字典类型状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dictionary.statusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/system/dict/{dictCode}/items": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "字典项"
                ],
                "summary": "按编码查询启用字典项",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典编码",
                        "name": "dictCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dictionary.ApiEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dictionary.ApiEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "dictionary.batchRequest": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "dictionary.dataRequest": {
            "type": "object",
            "properties": {
                "dictLabel": {
                    "type": "string"
                },
                "dictTypeId": {
                    "type": "integer"
                },
                "dictValue": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                }
            }
        },
        "dictionary.statusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "integer"
                }
            }
        },
        "dictionary.typeRequest": {
            "type": "object",
            "properties": {
                "dictCode": {
                    "type": "string"
                },
                "dictName": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}